enable fold gutters on a per language basis
